refactor: Update JWT cookie names in config.py
This commit updates the `config.py` file to include the `JWT_COOKIE_NAME` and `JWT_REFRESH_TOKEN_COOKIE_NAME` configuration variables. The `JWT_COOKIE_NAME` is now set to the value of the `JWT_COOKIE_NAME` environment variable or the default value of "access_token_cookie". Similarly, the `JWT_REFRESH_TOKEN_COOKIE_NAME` is set to the value of the `JWT_REFRESH_TOKEN_COOKIE_NAME` environment variable or the default value of "refresh_token_cookie". This change allows for easy customization of the cookie names for the access token and refresh token.