refactor: Update extensions.py to include Flask-Mail
This commit updates the `extensions.py` file to import and initialize the `Mail` extension from Flask-Mail. The `Mail` extension is used for sending emails in the application. By including this extension, we can now easily send emails using Flask-Mail, improving the functionality and user experience of the application.

Update `extensions.py` to include the `Mail` extension from Flask-Mail.